# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 1.0.0
  description: Enterprise Deal Commander (EDC) Phase 1 API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /v1/deals/{dealId}/gates:
    get:
      operationId: listGates
      tags: [gates]
      parameters:
        - { name: dealId, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: Gates
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GateListResponse"
  /v1/deals/{dealId}/gates/{gateCode}:
    put:
      operationId: updateGate
      tags: [gates]
      parameters:
        - { name: dealId, in: path, required: true, schema: { type: string } }
        - { name: gateCode, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GateUpdateInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GateMutationResponse"
  /v1/deals/{dealId}/gates/batch:
    put:
      operationId: updateGatesBatch
      tags: [gates]
      parameters:
        - { name: dealId, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GateBatchInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GateBatchResponse"

  # ============ BLOCKERS ============
...
  /v1/deals/{dealId}/alerts/{patternCode}/disposition:
    put:
      operationId: setDisposition
      tags: [governance]
      parameters:
        - { name: dealId, in: path, required: true, schema: { type: string } }
        - { name: patternCode, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DispositionInput"
      responses:
        "200":
          description: Disposition set
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DispositionResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: clearDisposition
      tags: [governance]
      parameters:
        - { name: dealId, in: path, required: true, schema: { type: string } }
        - { name: patternCode, in: path, required: true, schema: { type: string } }
      responses:
        "204":
          description: Cleared
  /v1/deals/{dealId}/interventions:
    post:
      operationId: launchIntervention
      tags: [governance]
      parameters:
        - { name: dealId, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InterventionInput"
      responses:
        "201":
          description: Launched
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InterventionResponse"
...
components:
  schemas:
...
    ErrorDetail:
      type: object
      properties:
        code: { type: string }
        message: { type: string }
        details:
          type: array
          items: { type: object, additionalProperties: true }
      required: [code, message]
    ErrorResponse:
      type: object
      properties:
        error: { $ref: "#/components/schemas/ErrorDetail" }
      required: [error]
...
    Gate:
      type: object
      properties:
        gateCode: { type: string }
        label: { type: string }
        gateGroup: { type: integer }
        description: { type: ["string", "null"] }
        sortOrder: { type: integer }
        isCompleted: { type: boolean }
        completedAt: { type: ["string", "null"] }
        completedBy: { type: ["string", "null"] }
        notes: { type: ["string", "null"] }
        prerequisiteGateCodes:
          type: array
          items: { type: string }
      required: [gateCode, label, gateGroup, sortOrder, isCompleted]
    IntegrityWarning:
      type: object
      properties:
        gateCode: { type: string }
        type: { type: string, enum: [out_of_order, regression] }
        message: { type: string }
      required: [gateCode, type, message]
    GateUpdateInput:
      type: object
      properties:
        is_completed: { type: boolean }
        notes: { type: string, maxLength: 1000 }
      required: [is_completed]
    GateBatchItem:
      type: object
      properties:
        gate_code: { type: string }
        is_completed: { type: boolean }
        notes: { type: string, maxLength: 1000 }
      required: [gate_code, is_completed]
    GateBatchInput:
      type: object
      properties:
        updates:
          type: array
          items: { $ref: "#/components/schemas/GateBatchItem" }
      required: [updates]
    GateListResponse:
      type: object
      properties:
        data:
          type: array
          items: { $ref: "#/components/schemas/Gate" }
      required: [data]
    GateMutationResponse:
      type: object
      properties:
        data: { $ref: "#/components/schemas/Gate" }
        integrityWarnings:
          type: array
          items: { $ref: "#/components/schemas/IntegrityWarning" }
      required: [data, integrityWarnings]
    GateBatchResponse:
      type: object
      properties:
        data:
          type: array
          items: { $ref: "#/components/schemas/Gate" }
        integrityWarnings:
          type: array
          items: { $ref: "#/components/schemas/IntegrityWarning" }
      required: [data, integrityWarnings]
...
    DispositionInput:
      type: object
      properties:
        disposition: { type: string, enum: [acknowledge, accept, snooze] }
        rationale: { type: string, minLength: 10, maxLength: 2000 }
        snooze_until_field_change: { type: string, minLength: 1, maxLength: 100 }
        snooze_duration_days: { type: integer, minimum: 1, maximum: 365 }
      required: [disposition]
    Disposition:
      type: object
      properties:
        id: { type: string }
        dealId: { type: string }
        patternCode: { type: string }
        disposition: { type: string }
        rationale: { type: ["string", "null"] }
        snoozeUntilFieldChange: { type: ["string", "null"] }
        snoozeUntil: { type: ["string", "null"] }
        createdBy: { type: string }
        createdAt: { type: string }
      required: [id, dealId, patternCode, disposition, createdBy]
    DispositionResponse:
      type: object
      properties:
        data: { $ref: "#/components/schemas/Disposition" }
      required: [data]
    InterventionInput:
      type: object
      properties:
        pattern_code: { type: string, minLength: 1, maxLength: 50 }
        checklist_id: { type: integer }
      required: [pattern_code, checklist_id]
    Intervention:
      type: object
      properties:
        id: { type: string }
        dealId: { type: string }
        patternCode: { type: string }
        checklistId: { type: integer }
        launchedBy: { type: string }
        launchedAt: { type: string }
      required: [id, dealId, patternCode, checklistId, launchedBy, launchedAt]
    InterventionResponse:
      type: object
      properties:
        data: { $ref: "#/components/schemas/Intervention" }
      required: [data]
...